pyproject.toml
[tool.poetry]
name = "ocr-fastapi"
version = "0.1.0"
description = "FastAPI backend for PaddleOCR with GPU on ECS"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic-settings = "^2.3.4"
python-multipart = "^0.0.9"
structlog = "^24.1.0"
pynvml = "^11.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.32.3"
tenacity = "^9.0.0"
Pillow = "^10.4.0"
numpy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
